Add test program. You can test whether it is answer.
Add another method to solve Q2 named t2.cpp.
This method use score to decide which pattern is the closest from answer.
And we use this pattern to mutate. Test if this mutation will gain score, we'll do it, otherwise don't.
The execution time about this is 21 seconds. It's faster than original one.